allow overriding access_grant strategy
